feat: add the letter dictionary in ASLRecognition
- Create a file Practice.kt for the letterDictionary as it is used in 2
  screens.
- Refactor the size with weigthed element to fit everything in the
  screen without needing to scroll.
- Refactor the screen UI to introduce the letterDictionary as requested
  by the coach to give exemple to the user when he tries to do some
  sign.
