+++ app/src/main/java/com/github/se/signify/ui/screens/home/ASLRecognition.kt
@@ -6,17 +6,18 @@ import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -24,6 +25,7 @@ import androidx.compose.ui.graphics.nativeCanvas
 import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.res.integerResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
@@ -33,8 +35,8 @@ import com.github.se.signify.model.hand.HandLandMarkViewModel
 import com.github.se.signify.model.navigation.NavigationActions
 import com.github.se.signify.ui.common.AnnexScreenScaffold
 import com.github.se.signify.ui.common.CameraBox
+import com.github.se.signify.ui.common.LetterDictionary
 import com.github.se.signify.ui.common.TextButton
-import com.google.mediapipe.tasks.components.containers.NormalizedLandmark
 
 /**
  * Composable that handles ASL recognition. It checks for camera permissions, launches the camera
@@ -47,18 +49,74 @@ fun ASLRecognition(
 ) {
   val buttonUriString = stringResource(id = R.string.button_uri_string_text)
   val context = LocalContext.current
+
+  val coroutineScope = rememberCoroutineScope()
+
+  val landmarksState = handLandMarkViewModel.landMarks().collectAsState()
+  val detectedGesture = handLandMarkViewModel.getSolution()
+
+  val imageResource =
+      if (landmarksState.value.isNullOrEmpty()) R.drawable.vector
+      else getIconResId(detectedGesture.first())
+
+  val makeASignText = stringResource(id = R.string.make_a_sign_text)
+  val displayText = if (landmarksState.value.isNullOrEmpty()) makeASignText else detectedGesture
+  val paintColor = MaterialTheme.colorScheme.background
   AnnexScreenScaffold(navigationActions = navigationActions, testTag = "ASLRecognitionScreen") {
     Box(
         modifier =
-            Modifier.fillMaxWidth()
-                .height(252.dp)
-                .background(MaterialTheme.colorScheme.background)) {
+            Modifier.fillMaxWidth().weight(1f).background(MaterialTheme.colorScheme.background)) {
           CameraBox(handLandMarkViewModel, "cameraPreview")
         }
 
     Spacer(modifier = Modifier.height(16.dp))
 
-    GestureOverlayView(handLandMarkViewModel)
+    LetterDictionary(
+        coroutineScope = coroutineScope,
+        numbOfHeaders = integerResource(R.integer.scroll_offset),
+        onClick = { _, _ -> })
+
+    Spacer(modifier = Modifier.height(16.dp))
+
+    // Draws the output text or the prompt to "Make a sign" on the screen.
+    Box(
+        modifier =
+            Modifier.fillMaxWidth()
+                .weight(1f)
+                .clip(RoundedCornerShape(10.dp))
+                .background(MaterialTheme.colorScheme.primary)
+                .border(
+                    width = 2.dp,
+                    color = MaterialTheme.colorScheme.primary,
+                    shape = RoundedCornerShape(10.dp)),
+        contentAlignment = Alignment.Center) {
+          Column(
+              modifier = Modifier.fillMaxSize().padding(16.dp),
+              horizontalAlignment = Alignment.CenterHorizontally) {
+                // Draws the output text or the prompt to "Make a sign" on the screen.
+                Canvas(modifier = Modifier.fillMaxSize().weight(1f).testTag("gestureOverlayView")) {
+                  val fontSize = 80f
+                  val textX = size.width * 0.5f
+                  val textY = (size.height / 2) + (fontSize / 3)
+                  drawContext.canvas.nativeCanvas.apply {
+                    val paint =
+                        android.graphics.Paint().apply {
+                          color = paintColor.toArgb()
+                          textSize = fontSize
+                          isAntiAlias = true
+                          textAlign = android.graphics.Paint.Align.CENTER
+                        }
+                    drawText(displayText, textX, textY, paint)
+                  }
+                }
+                // Displays the image associated with the detected ASL hand gesture
+                Icon(
+                    painter = painterResource(id = imageResource),
+                    contentDescription = "Letter gesture",
+                    tint = MaterialTheme.colorScheme.onPrimary,
+                    modifier = Modifier.fillMaxSize().weight(3f).testTag("handGestureImage"))
+              }
+        }
 
     Spacer(modifier = Modifier.height(16.dp))
     TextButton(
@@ -71,77 +129,4 @@ fun ASLRecognition(
         backgroundColor = MaterialTheme.colorScheme.primary,
         textColor = MaterialTheme.colorScheme.onPrimary)
   }
-}
-
-/**
- * Composable that overlays gesture detection on the UI. Displays landmarks and the recognized hand
- * gesture image.
- */
-@Composable
-fun GestureOverlayView(handLandMarkViewModel: HandLandMarkViewModel) {
-  val landmarksState = handLandMarkViewModel.landMarks().collectAsState()
-  val detectedGesture = handLandMarkViewModel.getSolution()
-
-  DrawnOutPut(landmarks = landmarksState.value, text = detectedGesture)
-  Spacer(modifier = Modifier.height(16.dp))
-  HandGestureImage(landmarks = landmarksState.value, gesture = detectedGesture)
-}
-/** Displays the image associated with the detected ASL hand gesture. */
-@Composable
-fun HandGestureImage(landmarks: List<NormalizedLandmark>?, gesture: String) {
-  val imageResource =
-      if (landmarks.isNullOrEmpty()) R.drawable.vector else getIconResId(gesture.first())
-  Box(
-      modifier =
-          Modifier.fillMaxWidth()
-              .height(200.dp)
-              .clip(RoundedCornerShape(size = 10.dp))
-              .background(MaterialTheme.colorScheme.primary)
-              .border(
-                  width = 2.dp,
-                  color = MaterialTheme.colorScheme.primary,
-                  shape = RoundedCornerShape(size = 10.dp))
-              .testTag("handGestureImage")) {
-        Icon(
-            painter = painterResource(id = imageResource),
-            contentDescription = "Letter gesture",
-            tint = MaterialTheme.colorScheme.onPrimary,
-            modifier = Modifier.size(200.dp).padding(16.dp).align(Alignment.Center))
-      }
-}
-
-/** Draws the output text or the prompt to "Make a sign" on the screen. */
-@Composable
-fun DrawnOutPut(landmarks: List<NormalizedLandmark>?, text: String) {
-  val makeASignText = stringResource(id = R.string.make_a_sign_text)
-  val displayText = if (landmarks.isNullOrEmpty()) makeASignText else text
-  val paintColor = MaterialTheme.colorScheme.background
-  Box(
-      modifier =
-          Modifier.fillMaxWidth()
-              .height(70.dp)
-              .clip(RoundedCornerShape(10.dp))
-              .background(MaterialTheme.colorScheme.primary)
-              .border(
-                  width = 2.dp,
-                  color = MaterialTheme.colorScheme.primary,
-                  shape = RoundedCornerShape(10.dp))
-              .testTag("gestureOverlayView"),
-      contentAlignment = Alignment.Center) {
-        Canvas(modifier = Modifier.fillMaxSize()) {
-          val fontSize = 80f
-          val textX = size.width * 0.5f
-          val textY = (size.height / 2) + (fontSize / 3)
-          drawContext.canvas.nativeCanvas.apply {
-            val paint =
-                android.graphics.Paint().apply {
-                  color = paintColor.toArgb()
-                  textSize = fontSize
-                  isAntiAlias = true
-                  textAlign = android.graphics.Paint.Align.CENTER
-                }
-            drawText(displayText, textX, textY, paint)
-          }
-        }
-      }
 }
